*** empty log message ***
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@1914 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/tb/sc4pd/source/sc4pd.hpp
@@ -102,12 +102,12 @@ class sc4pd_dsp
 	return 1 / Blocksize();
     }
     
-    inline float sc_filterloops()
+    inline int sc_filterloops()
     {
 	return Blocksize() / 3;
     }
     
-    inline float sc_filterremain()
+    inline int sc_filterremain()
     {
 	return Blocksize() % 3;
     }
